requirements.txt
dedalus-mcp>=0.1.0
dedalus-labs>=0.1.0
asyncpg>=0.29.0
httpx>=0.27.0
pandas>=2.2.0
supabase>=2.3.0
python-dotenv>=1.0.0
uvicorn>=0.27.0
pydantic>=2.0.0
